+++ 1-org/envs/shared/projects.tf
@@ -14,21 +14,29 @@
  * limitations under the License.
  */
 
+locals {
+  hub_and_spoke_roles = [
+    "roles/compute.instanceAdmin",
+    "roles/iam.serviceAccountAdmin",
+    "roles/resourcemanager.projectIamAdmin",
+    "roles/iam.serviceAccountUser",
+  ]
+}
+
 /******************************************
   Projects for log sinks
 *****************************************/
 
 module "org_audit_logs" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.0"
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-logging"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
-  activate_apis               = ["logging.googleapis.com", "bigquery.googleapis.com", "billingbudgets.googleapis.com"]
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-logging"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
+  activate_apis           = ["logging.googleapis.com", "bigquery.googleapis.com", "billingbudgets.googleapis.com"]
 
   labels = {
     environment       = "production"
@@ -45,16 +53,15 @@ module "org_audit_logs" {
 }
 
 module "org_billing_logs" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.0"
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-billing-logs"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
-  activate_apis               = ["logging.googleapis.com", "bigquery.googleapis.com", "billingbudgets.googleapis.com"]
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-billing-logs"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
+  activate_apis           = ["logging.googleapis.com", "bigquery.googleapis.com", "billingbudgets.googleapis.com"]
 
   labels = {
     environment       = "production"
@@ -75,16 +82,15 @@ module "org_billing_logs" {
 *****************************************/
 
 module "org_secrets" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-secrets"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
-  activate_apis               = ["logging.googleapis.com", "secretmanager.googleapis.com", "billingbudgets.googleapis.com"]
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-secrets"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
+  activate_apis           = ["logging.googleapis.com", "secretmanager.googleapis.com", "billingbudgets.googleapis.com"]
 
   labels = {
     environment       = "production"
@@ -105,16 +111,15 @@ module "org_secrets" {
 *****************************************/
 
 module "interconnect" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-interconnect"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
-  activate_apis               = ["billingbudgets.googleapis.com", "compute.googleapis.com"]
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-interconnect"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
+  activate_apis           = ["billingbudgets.googleapis.com", "compute.googleapis.com"]
 
   labels = {
     environment       = "production"
@@ -135,16 +140,15 @@ module "interconnect" {
 *****************************************/
 
 module "scc_notifications" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-scc"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
-  activate_apis               = ["logging.googleapis.com", "pubsub.googleapis.com", "securitycenter.googleapis.com", "billingbudgets.googleapis.com"]
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-scc"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
+  activate_apis           = ["logging.googleapis.com", "pubsub.googleapis.com", "securitycenter.googleapis.com", "billingbudgets.googleapis.com"]
 
   labels = {
     environment       = "production"
@@ -165,15 +169,14 @@ module "scc_notifications" {
 *****************************************/
 
 module "dns_hub" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-dns-hub"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-dns-hub"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
 
   activate_apis = [
     "compute.googleapis.com",
@@ -203,16 +206,15 @@ module "dns_hub" {
 *****************************************/
 
 module "base_network_hub" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  count                       = var.enable_hub_and_spoke ? 1 : 0
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-base-net-hub"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  count                   = var.enable_hub_and_spoke ? 1 : 0
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-base-net-hub"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
 
   activate_apis = [
     "compute.googleapis.com",
@@ -237,21 +239,28 @@ module "base_network_hub" {
   budget_amount               = var.base_net_hub_project_budget_amount
 }
 
+resource "google_project_iam_member" "network_sa_base" {
+  for_each = toset(var.enable_hub_and_spoke ? local.hub_and_spoke_roles : [])
+
+  project = module.base_network_hub[0].project_id
+  role    = each.key
+  member  = "serviceAccount:${var.networks_step_terraform_service_account_email}"
+}
+
 /******************************************
   Project for Restricted Network Hub
 *****************************************/
 
 module "restricted_network_hub" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  count                       = var.enable_hub_and_spoke ? 1 : 0
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  default_service_account     = "deprivilege"
-  name                        = "${var.project_prefix}-c-restricted-net-hub"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = google_folder.common.id
+  source                  = "terraform-google-modules/project-factory/google"
+  version                 = "~> 13.0"
+  count                   = var.enable_hub_and_spoke ? 1 : 0
+  random_project_id       = true
+  default_service_account = "deprivilege"
+  name                    = "${local.project_prefix}-c-restricted-net-hub"
+  org_id                  = local.org_id
+  billing_account         = local.billing_account
+  folder_id               = google_folder.common.id
 
   activate_apis = [
     "compute.googleapis.com",
@@ -274,4 +283,12 @@ module "restricted_network_hub" {
   budget_alert_pubsub_topic   = var.restricted_net_hub_project_alert_pubsub_topic
   budget_alert_spent_percents = var.restricted_net_hub_project_alert_spent_percents
   budget_amount               = var.restricted_net_hub_project_budget_amount
+}
+
+resource "google_project_iam_member" "network_sa_restricted" {
+  for_each = toset(var.enable_hub_and_spoke ? local.hub_and_spoke_roles : [])
+
+  project = module.restricted_network_hub[0].project_id
+  role    = each.key
+  member  = "serviceAccount:${var.networks_step_terraform_service_account_email}"
 }