+++ 2-environments/modules/env_baseline/monitoring.tf
@@ -20,14 +20,14 @@
 
 module "monitoring_project" {
   source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  random_project_id           = "true"
-  impersonate_service_account = var.terraform_service_account
-  name                        = "${var.project_prefix}-${var.environment_code}-monitoring"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
+  version                     = "~> 13.0"
+  random_project_id           = true
+  name                        = "${local.project_prefix}-${var.environment_code}-monitoring"
+  org_id                      = local.org_id
+  billing_account             = local.billing_account
   folder_id                   = google_folder.env.id
   disable_services_on_destroy = false
+  depends_on                  = [time_sleep.wait_30_seconds]
   activate_apis = [
     "logging.googleapis.com",
     "monitoring.googleapis.com",