+++ 4-projects/modules/single_project/main.tf
@@ -20,18 +20,17 @@ locals {
 }
 
 module "project" {
-  source                      = "terraform-google-modules/project-factory/google"
-  version                     = "~> 10.1"
-  random_project_id           = "true"
-  impersonate_service_account = var.impersonate_service_account
-  activate_apis               = distinct(concat(var.activate_apis, ["billingbudgets.googleapis.com"]))
-  name                        = "${var.project_prefix}-${var.business_code}-${local.env_code}-${var.project_suffix}"
-  org_id                      = var.org_id
-  billing_account             = var.billing_account
-  folder_id                   = var.folder_id
+  source            = "terraform-google-modules/project-factory/google"
+  version           = "~> 13.0"
+  random_project_id = true
+  activate_apis     = distinct(concat(var.activate_apis, ["billingbudgets.googleapis.com"]))
+  name              = "${var.project_prefix}-${var.business_code}-${local.env_code}-${var.project_suffix}"
+  org_id            = var.org_id
+  billing_account   = var.billing_account
+  folder_id         = var.folder_id
 
-  svpc_host_project_id = var.vpc_type == "" ? "" : data.google_compute_network.shared_vpc[0].project
-  shared_vpc_subnets   = var.vpc_type == "" ? [] : data.google_compute_network.shared_vpc[0].subnetworks_self_links # Optional: To enable subnetting, replace to "module.networking_project.subnetwork_self_link"
+  svpc_host_project_id = var.shared_vpc_host_project_id
+  shared_vpc_subnets   = var.shared_vpc_subnets # Optional: To enable subnetting, replace to "module.networking_project.subnetwork_self_link"
 
   vpc_service_control_attach_enabled = var.vpc_service_control_attach_enabled
   vpc_service_control_perimeter_name = var.vpc_service_control_perimeter_name